+++ src/main/java/de/jeff_media/ChestSort/ChestSortPlugin.java
@@ -68,7 +68,7 @@ public class ChestSortPlugin extends JavaPlugin implements de.jeff_media.ChestSo
 	String sortingMethod;
 	ArrayList<String> disabledWorlds;
 	ChestSortAPIHandler api;
-	final int currentConfigVersion = 35;
+	final int currentConfigVersion = 36;
 	boolean usingMatchingConfig = true;
 	protected boolean debug = false;
 	boolean verbose = true;