+++ src/main/java/de/jeff_media/ChestSort/hooks/CrateReloadedHook.java
@@ -1,15 +1,23 @@
 package de.jeff_media.ChestSort.hooks;
 
+import de.jeff_media.ChestSort.ChestSortPlugin;
 import org.bukkit.inventory.Inventory;
 import org.jetbrains.annotations.NotNull;
 
 public class CrateReloadedHook {
 
+    ChestSortPlugin main;
+
+    public CrateReloadedHook(ChestSortPlugin main) {
+        this.main=main;
+    }
+
     // CrateReloaded inventories seem to have a holder called cratereloaded.bo
     // Maybe this changes? We just check if the String starts with cratereloaded
-    public static boolean isCrate(Inventory inv) {
+    public boolean isCrate(Inventory inv) {
         if(inv==null) return false;
         if(inv.getHolder()==null) return false;
+        if(!main.getConfig().getBoolean("hook-cratereloaded",true)) return false;
         return inv.getHolder().getClass().getName().startsWith("cratereloaded");
     }