+++ src/main/java/de/jeff_media/ChestSort/hooks/HeadDatabaseHook.java
@@ -1,12 +1,20 @@
 package de.jeff_media.ChestSort.hooks;
 
+import de.jeff_media.ChestSort.ChestSortPlugin;
 import org.bukkit.inventory.Inventory;
 
 public class HeadDatabaseHook {
 
-    public static boolean isHeadDB(Inventory inv) {
+    ChestSortPlugin main;
+
+    public HeadDatabaseHook(ChestSortPlugin main) {
+        this.main=main;
+    }
+
+    public boolean isHeadDB(Inventory inv) {
         if(inv==null) return false;
         if(inv.getHolder() == null) return false;
+        if(!main.getConfig().getBoolean("hook-headdatabase",true)) return false;
         return inv.getHolder().getClass().getName().equals("me.arcaniax.hdb.object.HeadDatabaseHolder");
     }